remove _isAtBottom
