+++ apps/meteor/client/views/room/body/hooks/useListIsAtBottom.ts
@@ -1,3 +1,4 @@
+import { useMergedRefs } from '@rocket.chat/fuselage-hooks';
 import { useCallback, useRef } from 'react';
 
 import { isAtBottom } from '../../../../../app/ui/client/views/app/lib/scrolling';
@@ -6,10 +7,36 @@ import { withThrottling } from '../../../../../lib/utils/highOrderFunctions';
 export const useListIsAtBottom = () => {
 	const atBottomRef = useRef(true);
 
+	const innerBoxRef = useRef<HTMLDivElement | null>(null);
+
+	const sendToBottom = useCallback(() => {
+		innerBoxRef.current?.scrollTo({ left: 30, top: innerBoxRef.current?.scrollHeight });
+	}, []);
+
+	const sendToBottomIfNecessary = useCallback(() => {
+		if (atBottomRef.current === true) {
+			sendToBottom();
+		}
+	}, [atBottomRef, sendToBottom]);
+
 	const ref = useCallback((node: HTMLElement | null) => {
 		if (!node) {
 			return;
 		}
+
+		const messageList = node.querySelector('ul');
+
+		if (!messageList) {
+			return;
+		}
+
+		const observer = new ResizeObserver(() => {
+			if (atBottomRef.current === true) {
+				node.scrollTo({ left: 30, top: node.scrollHeight });
+			}
+		});
+
+		observer.observe(messageList);
 
 		node.addEventListener(
 			'scroll',
@@ -24,6 +51,8 @@ export const useListIsAtBottom = () => {
 
 	return {
 		atBottomRef,
-		innerRef: ref,
+		innerRef: useMergedRefs(ref, innerBoxRef) as unknown as React.MutableRefObject<HTMLDivElement | null>,
+		sendToBottom,
+		sendToBottomIfNecessary,
 	};
 };
